+++ dallinger/experiment_server/dashboard.py
@@ -6,11 +6,13 @@ from xml.sax.saxutils import escape
 
 import six
 import timeago
+from ansi2html import Ansi2HTMLConverter
 from flask import (
     Blueprint,
     abort,
     current_app,
     flash,
+    jsonify,
     redirect,
     render_template,
     request,
@@ -20,6 +22,7 @@ from flask.wrappers import Response
 from flask_login import UserMixin, current_user, login_required, login_user, logout_user
 from flask_login.utils import login_url as make_login_url
 from flask_wtf import FlaskForm
+from pygtail import Pygtail
 from six.moves.urllib.parse import urlencode
 from tzlocal import get_localzone
 from wtforms import BooleanField, HiddenField, PasswordField, StringField, SubmitField
@@ -28,9 +31,9 @@ from wtforms.validators import DataRequired, ValidationError
 import dallinger.db
 from dallinger import recruiters
 from dallinger.config import get_config
-from dallinger.db import get_all_mapped_classes
+from dallinger.db import get_all_mapped_classes, redis_conn
 from dallinger.heroku.tools import HerokuApp
-from dallinger.utils import deferred_route_decorator
+from dallinger.utils import deferred_route_decorator, get_logger_filename
 
 from .utils import date_handler, error_response, success_response
 
@@ -572,8 +575,6 @@ def mturk():
 @dashboard.route("/auto_recruit/<bool_val>", methods=["POST"])
 @login_required
 def auto_recruit(bool_val):
-    from dallinger.db import redis_conn
-
     num_val = int(bool_val)
     assert num_val in [0, 1]
     redis_conn.set("auto_recruit", num_val)
@@ -649,6 +650,116 @@ def lifecycle():
     return render_template(
         "dashboard_lifecycle.html", title="Experiment lifecycle Dashboard", **data
     )
+
+
+def clean_line_dict(line_dict, log_line_number):
+    msg = line_dict["message"]
+    if msg.endswith("-") and ("GET " in msg or " POST" in msg):
+        msg = '"'.join(msg.split('"')[1:])
+    msg = Ansi2HTMLConverter().convert(msg)
+    line_dict["message"] = msg
+    if log_line_number is not None:
+        line_dict["log_line_number"] = log_line_number
+    return line_dict
+
+
+line_number = redis_conn.get("line_number")
+
+
+LOG_FILE = get_logger_filename()
+
+
+def live_log():
+    """
+    Route that streams log file updates using SSE
+    """
+
+    def generate():
+        global line_number
+        line_number = (
+            int(line_number.decode("utf-8"))
+            if isinstance(line_number, bytes)
+            else (line_number or 0)
+        )
+        for line in Pygtail(LOG_FILE):
+            line_number += 1
+            redis_conn.set("line_number", line_number)
+            try:
+                line_dict = clean_line_dict(json.loads(line), line_number)
+                yield f"data:{json.dumps(line_dict)}\n\n"
+            except json.decoder.JSONDecodeError:
+                yield f"data:{line_number} | {line}\n\n"
+
+    return Response(generate(), mimetype="text/event-stream")
+
+
+def log_read_lines(line_range):
+    lines = []
+    max_line = max(line_range)
+    with open(LOG_FILE) as f:
+        for number, line in enumerate(f):
+            if number in line_range:
+                line_dict = clean_line_dict(json.loads(line), number)
+                lines.append(line_dict)
+            if number > max_line:
+                break
+    return lines
+
+
+def log_search_substring(substring):
+    with open(LOG_FILE) as f:
+        for number, line in enumerate(f):
+            if substring in line:
+                line_dict = clean_line_dict(json.loads(line), number)
+                yield f"data:{json.dumps(line_dict)}\n\n"
+    yield f"data:{json.dumps({'stop': True})}\n\n"
+
+
+@dashboard.route("/logs")
+@login_required
+def progress_log():
+    params = request.args
+    start, end = params.get("start", None), params.get("end", None)
+    n_nulled_params = sum(param is None for param in [start, end])
+
+    if query := params.get("query"):
+        return Response(log_search_substring(query), mimetype="text/event-stream")
+    if n_nulled_params == 2:
+        return live_log()
+    if n_nulled_params == 1:
+        return (
+            jsonify(
+                {
+                    "msg": "You should either provide both 'start' and 'end' or none of them."
+                }
+            ),
+            400,
+        )
+
+    start, end = int(start), int(end)
+    if start < 1:
+        return jsonify({"msg": "'start' must be greater than 0."}), 400
+    current_line_number = int(redis_conn.get("line_number"))
+    if end > current_line_number:
+        return (
+            jsonify(
+                {
+                    "msg": f"'end' must be less than the current line number {current_line_number}."
+                }
+            ),
+            400,
+        )
+    if start > end:
+        return (
+            jsonify(
+                {
+                    "msg": "The 'end' parameter should be less than the 'start' parameter."
+                }
+            ),
+            400,
+        )
+
+    return log_read_lines(range(start, end + 1))
 
 
 TABLE_DEFAULTS = {
